# apps/infrx-api/openapi/consumer.json
{
  "components": {
    "schemas": {
...
      "ApiKey": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "key_id": {
            "title": "Key Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "key_id",
          "name",
          "prefix",
          "created_at"
        ],
        "title": "ApiKey",
        "type": "object"
      },
...
      "CaptureRequest": {
        "additionalProperties": false,
        "description": "The key's mode plus the organization-level consent it is decided under.",
        "properties": {
          "consent_version": {
            "minimum": 0.0,
            "title": "Consent Version",
            "type": "integer"
          },
          "evaluation_consent": {
            "default": false,
            "title": "Evaluation Consent",
            "type": "boolean"
          },
          "mode": {
            "enum": [
              "off",
              "minimal",
              "full"
            ],
            "title": "Mode",
            "type": "string"
          },
          "retention_days": {
            "default": 30,
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          }
        },
        "required": [
          "mode",
          "consent_version"
        ],
        "title": "CaptureRequest",
        "type": "object"
      },
...
      "Consent": {
        "additionalProperties": false,
        "properties": {
          "effective_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective At"
          },
          "evaluation_consent": {
            "title": "Evaluation Consent",
            "type": "boolean"
          },
          "mode": {
            "enum": [
              "off",
              "minimal",
              "full"
            ],
            "title": "Mode",
            "type": "string"
          },
          "retention_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "mode",
          "retention_days",
          "evaluation_consent",
          "effective_at",
          "revoked_at"
        ],
        "title": "Consent",
        "type": "object"
      },
...
      "DataCategory": {
        "enum": [
          "request_content",
          "response_content",
          "media",
          "usage_metadata",
          "feedback"
        ],
        "title": "DataCategory",
        "type": "string"
      },
      "DataPurpose": {
        "description": "Four separate permissions (`01-architecture.md`). Capture implies none of\nthe other three, and a role is never consent.",
        "enum": [
          "capture",
          "provider_sharing",
          "external_judging",
          "training"
        ],
        "title": "DataPurpose",
        "type": "string"
      },
      "DataUseDoc": {
        "additionalProperties": false,
        "properties": {
          "consent": {
            "$ref": "#/components/schemas/Consent"
          },
          "grants": {
            "items": {
              "$ref": "#/components/schemas/Grant"
            },
            "title": "Grants",
            "type": "array"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/KeyCapture"
            },
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "consent",
          "keys",
          "grants"
        ],
        "title": "DataUseDoc",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field_errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Grant": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/DataCategory"
            },
            "title": "Categories",
            "type": "array"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "grant_id": {
            "title": "Grant Id",
            "type": "string"
          },
          "model_ids": {
            "items": {
              "type": "string"
            },
            "title": "Model Ids",
            "type": "array"
          },
          "provider_org_id": {
            "title": "Provider Org Id",
            "type": "string"
          },
          "purposes": {
            "items": {
              "$ref": "#/components/schemas/DataPurpose"
            },
            "title": "Purposes",
            "type": "array"
          },
          "retention_days": {
            "title": "Retention Days",
            "type": "integer"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "state": {
            "enum": [
              "active",
              "expired",
              "revoked"
            ],
            "title": "State",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "grant_id",
          "version",
          "provider_org_id",
          "model_ids",
          "categories",
          "purposes",
          "retention_days",
          "effective_at",
          "expires_at",
          "revoked_at",
          "state"
        ],
        "title": "Grant",
        "type": "object"
      },
...
      "GrantRequest": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/DataCategory"
            },
            "minItems": 1,
            "title": "Categories",
            "type": "array"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "grant_version": {
            "minimum": 0.0,
            "title": "Grant Version",
            "type": "integer"
          },
          "model_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Model Ids",
            "type": "array"
          },
          "provider_org_id": {
            "format": "uuid",
            "title": "Provider Org Id",
            "type": "string"
          },
          "purposes": {
            "items": {
              "$ref": "#/components/schemas/DataPurpose"
            },
            "minItems": 1,
            "title": "Purposes",
            "type": "array"
          },
          "retention_days": {
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          }
        },
        "required": [
          "provider_org_id",
          "model_ids",
          "categories",
          "purposes",
          "retention_days",
          "grant_version"
        ],
        "title": "GrantRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeyCapture": {
        "additionalProperties": false,
        "properties": {
          "effective_mode": {
            "enum": [
              "off",
              "minimal",
              "full"
            ],
            "title": "Effective Mode",
            "type": "string"
          },
          "key_id": {
            "title": "Key Id",
            "type": "string"
          },
          "mode": {
            "enum": [
              "off",
              "minimal",
              "full"
            ],
            "title": "Mode",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key_id",
          "name",
          "mode",
          "effective_mode"
        ],
        "title": "KeyCapture",
        "type": "object"
      },
      "KeyCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "KeyCreate",
        "type": "object"
      },
      "KeyCreated": {
        "additionalProperties": false,
        "description": "`secret` is set only on the call that created the row; a replay carries null and\n`secret_returned: false` - a lost secret is a revoke and a new key, never a re-reveal.",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/ApiKey"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          },
          "secret_returned": {
            "title": "Secret Returned",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "secret",
          "secret_returned",
          "replayed"
        ],
        "title": "KeyCreated",
        "type": "object"
      },
...
      "KeySummary": {
        "additionalProperties": false,
        "description": "Never the secret or its hash.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "created_at",
          "last_used_at",
          "revoked_at"
        ],
        "title": "KeySummary",
        "type": "object"
      },
...
      "ListPage_Grant_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Grant"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "ListPage[Grant]",
        "type": "object"
      },
      "ListPage_KeySummary_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/KeySummary"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "ListPage[KeySummary]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "bearerFormat": "sk-infrx",
        "description": "A consumer or provider-dev API key (`/v1/*`).",
        "scheme": "bearer",
        "type": "http"
      },
      "CheckpointSignature": {
        "description": "HMAC of the raw body by a LAB_CHECKPOINT_KEYS key.",
        "in": "header",
        "name": "X-Infrx-Signature",
        "type": "apiKey"
      },
      "OperatorKey": {
        "bearerFormat": "sk-infrx",
        "description": "An operator-audience API key.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionBearer": {
        "bearerFormat": "JWT",
        "description": "The signed-in user's Supabase session, forwarded by the web app's server; verified by the auth server.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "infrx consumer",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/console/v1/data-grants": {
      "get": {
        "operationId": "getConsoleV1DataGrants",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_Grant_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "List Grants"
      },
      "post": {
        "operationId": "postConsoleV1DataGrants",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Grant"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Put Grant"
      }
    },
    "/console/v1/data-grants/{grant_id}": {
      "delete": {
        "operationId": "deleteConsoleV1DataGrantsByGrantId",
        "parameters": [
          {
            "in": "path",
            "name": "grant_id",
            "required": true,
            "schema": {
              "title": "Grant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Grant"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Revoke Grant"
      }
    },
    "/console/v1/data-use": {
      "get": {
        "operationId": "getConsoleV1DataUse",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataUseDoc"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Read Data Use"
      }
    },
    "/console/v1/keys": {
      "get": {
        "operationId": "getConsoleV1Keys",
        "parameters": [
          {
            "description": "page size 1..100 (default 25)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "page size 1..100 (default 25)",
              "title": "Limit"
            }
          },
          {
            "description": "the next_cursor of the previous page, passed back as given",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "the next_cursor of the previous page, passed back as given",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_KeySummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Handler"
      },
      "post": {
        "operationId": "postConsoleV1Keys",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Create Key"
      }
    },
    "/console/v1/keys/{key_id}": {
      "delete": {
        "operationId": "deleteConsoleV1KeysByKeyId",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKey"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Revoke Key"
      }
    },
    "/console/v1/keys/{key_id}/capture": {
      "put": {
        "operationId": "putConsoleV1KeysByKeyIdCapture",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataUseDoc"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Put Capture"
      }
    },
...
  }
}